Ability to search messages, Messages modal and message input refactoring
