+++ src/components/pages/Donate.js
@@ -5,13 +5,14 @@ import WAValidator from 'multicoin-address-validator';
 import { AppContext } from '../ContextProvider';
 import { maskAddress } from '../../helpers/utils';
 import { useFormInput, useFormValidation } from '../../helpers/hooks';
+import FormLabelDescription from '../elements/FormLabelDescription';
 
 
 const Donate = props => {
   const { actions, state } = useContext(AppContext);
   const { createWallet, sendTx } = actions;
   const { appSettings, layout, marketData, userSettings, wallets } = state;
-  const { coinDecimals, defaultFee, messageFee, feePerChar } = appSettings;
+  const { coinDecimals, defaultFee, messageFee, messageLimit, feePerChar } = appSettings;
   const { formSubmitted, sendTxResponse, walletsLoaded } = layout;
 
   const params = new URLSearchParams(props.location.search);
@@ -133,14 +134,20 @@ const Donate = props => {
             >
               <div className="row donateData">
                 <div className="col-lg-12">
-                  <div className="form-layout form-layout-7">
+                  <div className="form-layout form-layout-7 donate-form">
                     <div className="row no-gutters">
-                      <div className="col-5 col-sm-2">Pay to</div>
+                      <div className="col-5 col-sm-2">
+                        Pay to
+                        <FormLabelDescription>Receiver's address</FormLabelDescription>
+                      </div>
                       <div
                         className="col-7 col-sm-10 wallet-address">{address} {recipientName && `(${recipientName})`}</div>
                     </div>
                     <div className="row no-gutters">
-                      <div className="col-5 col-sm-2">From Wallet</div>
+                      <div className="col-5 col-sm-2">
+                        From Wallet
+                        <FormLabelDescription>Your wallet from which funds will be sent</FormLabelDescription>
+                      </div>
                       <div className="col-7 col-sm-10">
                         {walletsLoaded && Object.keys(availableWallets).length > 0 &&
                         <select
@@ -174,7 +181,10 @@ const Donate = props => {
                       </div>
                     </div>
                     <div className="row no-gutters">
-                      <div className="col-5 col-sm-2">Amount</div>
+                      <div className="col-5 col-sm-2">
+                        Amount
+                        <FormLabelDescription>Amount of CCX to send</FormLabelDescription>
+                      </div>
                       <div className="col-7 col-sm-10">
                         <input
                           {...bindAmount}
@@ -190,57 +200,78 @@ const Donate = props => {
                         />
                         <div className="float-left mg-l-10">
                           BTC: {btcValue.toLocaleString(undefined, btcFormatOptions)}<br />
-						  USD: {usdValue.toLocaleString(undefined, usdFormatOptions)}
+						              USD: {usdValue.toLocaleString(undefined, usdFormatOptions)}
                         </div>
                       </div>
                     </div>
                     <div className="row no-gutters">
-                      <div className="col-5 col-sm-2">Message</div>
+                      <div className="col-5 col-sm-2">
+                        Message
+                        <FormLabelDescription>Optional message to include in this transaction</FormLabelDescription>
+                      </div>
                       <div className="col-7 col-sm-10">
-                        <input
-                          {...bindMessage}
-                          size={6}
-                          className="form-control maxWidth"
-                          placeholder="Message"
-                          name="message"
-                          type="text"
-                          disabled={Object.keys(availableWallets).length === 0}
-                        />
+                        <div className="input-group">
+                          <input
+                            {...bindMessage}
+                            size={6}
+                            className="form-control"
+                            placeholder="Message"
+                            name="message"
+                            type="text"
+                            maxLength={messageLimit}
+                            disabled={Object.keys(availableWallets).length === 0}
+                          />
+                          <div className="input-group-append">
+                            <span className="input-group-text">
+                              <small>
+                                <strong>
+                                  {message.length}/{messageLimit} Characters
+                                </strong>
+                              </small>
+                            </span>
+                          </div>
+                        </div>
                       </div>
                     </div>
                     <div className="row no-gutters">
                       {userSettings.twoFAEnabled
                         ? <>
-                          <div className="col-5 col-sm-2">2FA Code</div>
-                          <div className="col-7 col-sm-10">
-                            <input
-                              {...bindTwoFACode}
-                              size={6}
-                              placeholder="2 Factor Authentication"
-                              className="form-control autoWidth"
-                              name="twoFACode"
-                              type="number"
-                              minLength={6}
-                              maxLength={6}
-                              disabled={Object.keys(availableWallets).length === 0}
-                            />
-                          </div>
-                        </>
+                            <div className="col-5 col-sm-2">
+                              2FA Code
+                              <FormLabelDescription>2 Factor Authentication code</FormLabelDescription>
+                            </div>
+                            <div className="col-7 col-sm-10">
+                              <input
+                                {...bindTwoFACode}
+                                size={6}
+                                placeholder="2 Factor Authentication"
+                                className="form-control autoWidth"
+                                name="twoFACode"
+                                type="number"
+                                minLength={6}
+                                maxLength={6}
+                                disabled={Object.keys(availableWallets).length === 0}
+                              />
+                            </div>
+                          </>
                         : <>
-                          <div className="col-5 col-sm-2">Password</div>
-                          <div className="col-7 col-sm-10">
-                            <input
-                              {...bindPassword}
-                              size={6}
-                              className="form-control"
-                              placeholder="Password"
-                              name="password"
-                              type="password"
-                              minLength={8}
-                              disabled={Object.keys(availableWallets).length === 0}
-                            />
-                          </div>
-                        </>
+                            <div className="col-5 col-sm-2">
+                              Password
+                              <FormLabelDescription>Your password</FormLabelDescription>
+                            </div>
+                            <div className="col-7 col-sm-10">
+                              <input
+                                {...bindPassword}
+                                size={6}
+                                className="form-control"
+                                placeholder="Password"
+                                name="password"
+                                type="password"
+                                minLength={8}
+                                disabled={Object.keys(availableWallets).length === 0}
+                              />
+                            </div>
+                          </>
                       }
                     </div>
                   </div>
@@ -250,7 +281,7 @@ const Donate = props => {
               <div>
                 <button
                   type="submit"
-                  className="btn btn-outline-primary btn-uppercase-sm"
+                  className={`btn btn-uppercase-sm btn-send ${formValid ? 'btn-outline-success' : 'btn-outline-danger'}`}
                   disabled={formSubmitted || !formValid}
                 >
                   SEND
@@ -263,15 +294,15 @@ const Donate = props => {
                   sendTxResponse.status === 'error'
                     ? <div className="text-danger">{sendTxResponse.message}</div>
                     : <>
-                      TX Hash: <a
-                      href={`${appSettings.explorerURL}/index.html?hash=${sendTxResponse.message.transactionHash}#blockchain_transaction`}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                    >
-                      {sendTxResponse.message.transactionHash}
-                    </a><br/>
-                      Secret Key: {sendTxResponse.message.transactionSecretKey}
-                    </>
+                        TX Hash: <a
+                        href={`${appSettings.explorerURL}/index.html?hash=${sendTxResponse.message.transactionHash}#blockchain_transaction`}
+                        target="_blank"
+                        rel="noopener noreferrer"
+                      >
+                        {sendTxResponse.message.transactionHash}
+                      </a><br/>
+                        Secret Key: {sendTxResponse.message.transactionSecretKey}
+                      </>
                 }
               </div>
               }
